feat: Integer更改為Long

+++ src/main/java/com/e_sun_interview/vote/repository/CMSRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CMSRepository extends JpaRepository<Voting_items, Integer> {
+public interface CMSRepository extends JpaRepository<Voting_items, Long> {
 }
